// src/api/generated/api.ts
export interface ExperienceFlat {
    'id'?: number;
    'title': string;
    'function': string;
    'description': string;
    'initial_date'?: string;
    'end_date'?: string;
    'actual'?: boolean;
    'skills'?: Array<SkillFlat>;
}
// ...
export interface SkillFlat {
    'id': number;
    'title': string;
    'revelance'?: number;
    'description'?: string;
}